+++ src/lib/runtime.ts
@@ -1,8 +1,16 @@
-import {StorageSyncKey, chromeStorageGet} from "./chromeStorage";
 import {Logger} from "./logger";
 import Bluebird from "bluebird";
+import {XMTP_CONVERSATION_FLAG} from "../types/wallet/messages";
+import {StorageSyncKey, chromeStorageGet} from "./chromeStorage";
+import rgbHex from "rgb-hex";
+import {sleep} from "./wallet/sleep";
 
-const permissions = ["contextMenus", "notifications", "sidePanel", "tabs"];
+export enum BadgeColor {
+  Blue = "#1976d2",
+  Green = "#4caf50",
+}
+
+const permissions = ["contextMenus", "notifications", "tabs"];
 
 export const hasOptionalPermissions = async (): Promise<boolean> => {
   return await chrome.permissions.contains({permissions});
@@ -76,52 +84,66 @@ export const setIcon = async (
   });
 };
 
-export const incrementBadgeCount = async (
-  color: "blue" | "green" = "green",
-) => {
+export const getBadgeCount = async (color: BadgeColor) => {
+  // determine if color matches query
+  const currentRgb = await chrome.action.getBadgeBackgroundColor({});
+  const currentHex = rgbHex(currentRgb[0], currentRgb[1], currentRgb[2]);
+  if (!color.toLowerCase().includes(currentHex.toLowerCase())) {
+    return 0;
+  }
+
+  // get current count
   let currentCount = await chrome.action.getBadgeText({});
   if (!currentCount) {
     currentCount = "0";
   }
   try {
-    await setBadgeCount(parseInt(currentCount) + 1, color);
+    return parseInt(currentCount);
   } catch (e) {
-    Logger.warn("unable to increment badge count", e);
+    // ignore error
   }
+  return 0;
+};
+
+export const incrementBadgeCount = async (color: BadgeColor) => {
+  const currentCount = await getBadgeCount(color);
+  await setBadgeCount(currentCount + 1, color);
 };
 
 export const setBadgeCount = async (
   count: number,
-  color: "blue" | "green" = "green",
+  color: BadgeColor = BadgeColor.Green,
 ) => {
-  await chrome.action.setBadgeBackgroundColor({
-    color: color === "green" ? "#4caf50" : "#1976d2",
-  });
+  await chrome.action.setBadgeBackgroundColor({color});
   await chrome.action.setBadgeTextColor({color: "#ffffff"});
   await chrome.action.setBadgeText({text: count > 0 ? String(count) : ""});
 };
 
-export const getAllPopups = (): Window[] => {
-  return chrome.extension.getViews({type: "tab"});
-};
-
-export const focusAllPopups = async () => {
-  await Promise.all([focusExtensionPopups(), focusKnownPopup()]);
-};
-
-export const focusExtensionPopups = async () => {
+export const focusExtensionWindows = async (
+  windowType?: chrome.tabs.QueryInfo["windowType"],
+) => {
+  let focussedCount = 0;
   try {
-    const popupTabs = await chrome.tabs.query({windowType: "popup"});
-    const allWindows = getAllPopups();
-    await Bluebird.map(allWindows, async (window) => {
-      const tab = popupTabs.find((t) => t.url?.includes(window.location.href));
-      if (tab) {
+    // retrieve all active windows for requested type
+    const allOpenWindows = await chrome.tabs.query({windowType});
+    const extensionBaseUrl = chrome.runtime.getURL("");
+
+    // focus for side panels
+    await Bluebird.map(allOpenWindows, async (tab) => {
+      if (tab.url?.includes(extensionBaseUrl)) {
         await chrome.windows.update(tab.windowId, {focused: true});
+        focussedCount++;
       }
     });
   } catch (e) {
     // ignore error
   }
+
+  // if no popup was found to focus, try alternative method
+  if (focussedCount === 0) {
+    return await focusKnownPopup();
+  }
+  return focussedCount;
 };
 
 export const focusKnownPopup = async () => {
@@ -130,13 +152,14 @@ export const focusKnownPopup = async () => {
       StorageSyncKey.WindowId,
       "session",
     );
-    if (!windowId) {
-      return;
+    if (windowId) {
+      await chrome.windows.update(windowId, {focused: true});
+      return 1;
     }
-    await chrome.windows.update(windowId, {focused: true});
   } catch (e) {
     // ignore error
   }
+  return 0;
 };
 
 export const createNotification = async (
@@ -147,14 +170,62 @@ export const createNotification = async (
   priority?: number,
 ) => {
   if (chrome.notifications) {
-    chrome.notifications.create(id, {
-      type: "basic",
-      title,
-      iconUrl: chrome.runtime.getURL("/icon/128.png"),
-      message,
-      isClickable: true,
-      contextMessage,
-      priority,
-    });
+    // a callback to determine when the notification is finished
+    let isComplete = false;
+    const callback = () => {
+      isComplete = true;
+    };
+
+    // request the notification to be created
+    chrome.notifications.create(
+      id,
+      {
+        type: "basic",
+        title,
+        iconUrl: chrome.runtime.getURL("/icon/128.png"),
+        message,
+        isClickable: true,
+        contextMessage,
+        priority,
+      },
+      callback,
+    );
+
+    // wait for complete
+    while (!isComplete) {
+      await sleep(250);
+    }
   }
+};
+
+export const openSidePanel = async (opts?: {
+  address?: string;
+  windowId?: number;
+}) => {
+  if (chrome.sidePanel) {
+    try {
+      // determine the current window ID
+      const windowId =
+        opts?.windowId || (await chrome.windows.getCurrent())?.id;
+      if (!windowId) {
+        return;
+      }
+
+      // build the URL used to open the side panel
+      const sidePanelUrl = `${chrome.runtime.getURL(`index.html${opts?.address ? `?${XMTP_CONVERSATION_FLAG}=${opts.address}` : ""}`)}#messages`;
+
+      // open the side panel and set the URL concurrently, to avoid a timing
+      // bug in the chrome user gesture flag handling
+      await Promise.all([
+        chrome.sidePanel.open({windowId}),
+        chrome.sidePanel.setOptions({enabled: true, path: sidePanelUrl}),
+      ]);
+      return true;
+    } catch (e) {
+      // gracefully handle the error and fallback to opening the chat within
+      // the same window instead of side panel
+      Logger.warn(e, "Popup", "Unable to open message side panel");
+    }
+  }
+  return false;
 };