+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.1.31
+* Update XMTP version to 12.1.0
+* Enhanced message spam protections
+* Open messages in browser side panel
+* Messages context menu
+
+## 3.1.28
+* Reduce the set of required extension permissions
+* Make permissions optional where possible
+
 ## 3.1.25
 * Support for Unstoppable Lite Wallet, a wallet for domainers
 * Support for Unstoppable Messaging, powered by XMTP